refactor: refactor and adjust suggestion weights
