+++ cspell.schema.json
@@ -30,6 +30,32 @@
       ],
       "type": "string"
     },
+    "CharacterSet": {
+      "description": "This is a set of characters that can include `-` or `|`\n- `-` - indicates a range of characters: `a-c` => `abc`\n- `|` - is a group separator, indicating that the characters on either side    are not related.",
+      "type": "string"
+    },
+    "CharacterSetCosts": {
+      "additionalProperties": false,
+      "properties": {
+        "characters": {
+          "$ref": "#/definitions/CharacterSet",
+          "description": "This is a set of characters that can include `-` or `|`\n- `-` - indicates a range of characters: `a-c` => `abc`\n- `|` - is a group separator, indicating that the characters on either side    are not related."
+        },
+        "cost": {
+          "description": "the cost to insert / delete / replace / swap the characters in a group",
+          "type": "number"
+        },
+        "penalty": {
+          "description": "The penalty cost to apply if the accent is used. This is used to discourage",
+          "type": "number"
+        }
+      },
+      "required": [
+        "characters",
+        "cost"
+      ],
+      "type": "object"
+    },
     "CostMapDefInsDel": {
       "additionalProperties": false,
       "properties": {
@@ -365,6 +391,40 @@
       "additionalProperties": false,
       "description": "Use by dictionary authors to help improve the quality of suggestions given from the dictionary.\n\nAdded with `v5.16.0`.",
       "properties": {
+        "accents": {
+          "anyOf": [
+            {
+              "$ref": "#/definitions/CharacterSet"
+            },
+            {
+              "items": {
+                "$ref": "#/definitions/CharacterSetCosts"
+              },
+              "type": "array"
+            }
+          ],
+          "default": "̀-́",
+          "description": "The accent characters"
+        },
+        "alphabet": {
+          "anyOf": [
+            {
+              "$ref": "#/definitions/CharacterSet"
+            },
+            {
+              "items": {
+                "$ref": "#/definitions/CharacterSetCosts"
+              },
+              "type": "array"
+            }
+          ],
+          "default": "a-zA-Z",
+          "description": "The alphabet to use."
+        },
+        "costs": {
+          "$ref": "#/definitions/EditCosts",
+          "description": "Define edit costs."
+        },
         "hunspellInformation": {
           "$ref": "#/definitions/HunspellInformation",
           "description": "Used by dictionary authors"
@@ -404,6 +464,37 @@
         }
       ],
       "description": "Reference to a dictionary by name. One of:\n-  {@link  DictionaryRef } \n-  {@link  DictionaryNegRef }"
+    },
+    "EditCosts": {
+      "additionalProperties": false,
+      "properties": {
+        "accentCosts": {
+          "default": 1,
+          "description": "The cost to add / remove an accent This should be very cheap, it helps with fixing accent issues.",
+          "type": "number"
+        },
+        "baseCost": {
+          "default": 100,
+          "description": "This is the base cost for making an edit.",
+          "type": "number"
+        },
+        "capsCosts": {
+          "default": 1,
+          "description": "The cost to change capitalization. This should be very cheap, it helps with fixing capitalization issues.",
+          "type": "number"
+        },
+        "firstLetterPenalty": {
+          "default": 4,
+          "description": "The extra cost incurred for changing the first letter of a word. This value should be less than `100 - baseCost`.",
+          "type": "number"
+        },
+        "nonAlphabetCosts": {
+          "default": 110,
+          "description": "This is the cost for characters not in the alphabet.",
+          "type": "number"
+        }
+      },
+      "type": "object"
     },
     "FSPathResolvable": {
       "$ref": "#/definitions/FsPath",
@@ -447,6 +538,11 @@
               "description": "The cost to add / remove an accent This should be very cheap, it helps with fixing accent issues.",
               "type": "number"
             },
+            "baseCost": {
+              "default": 100,
+              "description": "This is the base cost for making an edit.",
+              "type": "number"
+            },
             "capsCosts": {
               "default": 1,
               "description": "The cost to change capitalization. This should be very cheap, it helps with fixing capitalization issues.",
@@ -454,7 +550,7 @@
             },
             "firstLetterPenalty": {
               "default": 4,
-              "description": "The extra cost incurred for changing the first letter of a word. This value should be less than `100 - tryCharCost`.",
+              "description": "The extra cost incurred for changing the first letter of a word. This value should be less than `100 - baseCost`.",
               "type": "number"
             },
             "ioConvertCost": {
@@ -463,7 +559,7 @@
               "type": "number"
             },
             "keyboardCost": {
-              "default": 94,
+              "default": 99,
               "description": "The cost of replacing or swapping any adjacent keyboard characters.\n\nThis should be slightly cheaper than `tryCharCost`.",
               "type": "number"
             },
@@ -472,14 +568,18 @@
               "description": "mapSet replacement cost is the cost to substitute one character with another from the same set.\n\nMap characters are considered very similar to each other and are often the cause of simple mistakes.",
               "type": "number"
             },
+            "nonAlphabetCosts": {
+              "default": 110,
+              "description": "This is the cost for characters not in the alphabet.",
+              "type": "number"
+            },
             "replaceCosts": {
               "default": 75,
               "description": "The cost to substitute pairs found in the replace settings.",
               "type": "number"
             },
             "tryCharCost": {
-              "default": 95,
-              "description": "The cost of inserting / deleting / or swapping any `tryChars`",
+              "description": "The cost of inserting / deleting / or swapping any `tryChars` Defaults to `baseCosts`",
               "type": "number"
             }
           },
