+++ packages/cspell-trie-lib/src/lib/mappers/mapCosts.ts
@@ -1,0 +1,30 @@
+import { HunspellCosts, EditCosts } from '../models/DictionaryInformation';
+import { clean } from '../utils/util';
+
+export type EditCostsRequired = Required<EditCosts>;
+export type HunspellCostsRequired = Required<HunspellCosts>;
+
+const defaultEditCosts: EditCostsRequired = {
+    accentCosts: 1,
+    baseCost: 100,
+    capsCosts: 1,
+    firstLetterPenalty: 4,
+    nonAlphabetCosts: 110,
+};
+
+const defaultHunspellCosts: HunspellCostsRequired = {
+    ...defaultEditCosts,
+    ioConvertCost: 30,
+    keyboardCost: 99,
+    mapCost: 25,
+    replaceCosts: 75,
+    tryCharCost: 100,
+};
+
+export function mapHunspellCosts(costs: HunspellCosts = {}): HunspellCostsRequired {
+    return { ...defaultHunspellCosts, ...clean({ ...costs }) };
+}
+
+export function mapEditCosts(costs: EditCosts = {}): EditCostsRequired {
+    return { ...defaultEditCosts, ...clean({ ...costs }) };
+}
