+++ packages/cspell-trie-lib/src/lib/utils/util.ts
@@ -2,6 +2,12 @@ export function isDefined<T>(a: T | undefined): a is T {
     return a !== undefined;
 }
 
+/**
+ * Remove any fields with an `undefined` value.
+ * **MODIFIES THE OBJECT**
+ * @param t - object to clean
+ * @returns t
+ */
 export function clean<T>(t: Partial<T>): Partial<T> {
     const r: Partial<T> = {};
     for (const k of Object.keys(t) as (keyof T)[]) {
