+++ packages/cspell-types/src/index.ts
@@ -3,6 +3,7 @@ export { ConfigFields } from './configFields';
 export type { CSpellUserSettingsFields } from './configFields';
 export * from './CSpellReporter';
 export * from './CSpellSettingsDef';
+export type { CharacterSet, CharacterSetCosts, DictionaryInformation, EditCosts } from './DictionaryInformation';
+export type { Feature, Features } from './features';
 export type { SuggestionCostMapDef, SuggestionCostsDefs } from './suggestionCostsDef';
 export type { TextDocumentOffset, TextOffset } from './TextOffset';
-export type { Features, Feature } from './features';
